Fix ordering bug in proxyHandler.endpointMaps
Since maps don't have a guaranteed return order in golang coupled with
route fitness determined by matching just the prefix of the
request.URL.path with endpointMap.path, it was possible for the
incorrect path to followed if they were not returned in the expected
order.

Scenario:
If endpointMaps contained a route which matched the paths `/` and
`/foo`, a request for `/foobar` would satisfy both paths and would be
decided by which routeMap is returned from the map structure.

Instead, replace with arrays which have deterministic order and allows
the developer to control priority of the used endpoints.